% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addLabel.R
\name{addLabel}
\alias{addLabel}
\title{Add label for the leaflet map}
\usage{
addLabel(data)
}
\arguments{
\item{data}{Dataframe that contain Country/Region, confirmed, deceased and recovered columns}
}
\value{
New column named "label" is created in the data table.
}
\description{
This function create HTML label for the types of cases. When mouse is hovered
on leaflet map, the label will appear and give addition details.
}
