% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_total}
\alias{data_total}
\title{COVID19 data from 22/01/2020 to 06/10/2020}
\format{
A tibble with 48692 observations and 7 variables
\itemize{
\item \strong{Country/Region}: Country/ Region name
\item \strong{date}: The date of observation, using \code{Date} class.
\item \strong{lat}: The latitude code.
\item \strong{long}: The longitude code.
\item \strong{confirmed}: The number of confirmed cases.
\item \strong{deceased}: The number of deceased cases.
\item \strong{recovered}: The number of recovered cases.
}
}
\source{
\href{https://github.com/CSSEGISandData/COVID-19}{John Hopkins Github}
}
\usage{
data_total
}
\description{
A dataset contains the number of COVID19 cases worldwide from 22/01/2020 to 06/10/2020 as well as the date and the region in which cases were recored.
There are 3 type of cases: Confirmed, deceased and recovered
}
\keyword{datasets}
