% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_total_newdaily}
\alias{data_total_newdaily}
\title{Data for new cases of COVID19 from 22/01/2020 to 06/10/2020}
\format{
A tibble with 292152 observations and 6 variables
\itemize{
\item \strong{Country/Region}: Country/ Region name
\item \strong{date}: The date of observation, using \code{Date} class.
\item \strong{lat}: The latitude code.
\item \strong{long}: The longitude code.
\item \strong{var}: The cases type, \code{c("Confirmed cases","Recovered cases","Deceased cases", "New daily cases","New death cases","New recovered cases")}
\item \strong{value} - The number of cases
}
}
\source{
\href{https://github.com/CSSEGISandData/COVID-19}{John Hopkins Github}
}
\usage{
data_total_newdaily
}
\description{
#'#'A dataset contains the number of COVID19 cases worldwide from 22/01/2020 to 06/10/2020.
There are 6 type of cases: Confirmed, deceased, recovered, new daily confirmed,
new daily death and new daily recovered
}
\keyword{datasets}
