% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSummaryDT.R
\name{getSummaryDT}
\alias{getSummaryDT}
\title{Function to draw the summary table}
\usage{
getSummaryDT(data, groupBy)
}
\arguments{
\item{data}{Data need to be data_total}

\item{groupBy}{the variables that user want to group and summarise.}
}
\value{
A summary data table will be created.
}
\description{
This function create a summary data table that shows the detailed
number of confirmed, deceased and Estimated recovered cases.
}
\examples{
getSummaryDT(data_total, date)
}
