% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sliderDateinput.R
\name{sliderDateinput}
\alias{sliderDateinput}
\title{This function create the slider input in shiny app}
\usage{
sliderDateinput(inputId, label = "Select date", x, ...)
}
\arguments{
\item{inputId}{The input slot that will be used to access the value.}

\item{x}{The variables that will be used to construct the range of value}

\item{...}{Arguments passed to sliderInput().}
}
\value{
Export a slider input in Shiny app.
}
\description{
Constructs a slider widget to select a date value from a range of value.
}
\details{
Title
}
\examples{

library(shiny)
ui <-   fluidPage(
 fluidRow(
   column(
     sliderDateinput(inputId  = "timeSlider",x = data_total$date),
     width = 12)
 )
)
server <- function(input, output) {
}
shinyApp(ui = ui, server = server)
}
