% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariseData.R
\name{summariseData}
\alias{summariseData}
\alias{summarizeData}
\title{Summarise data to fewer rows}
\usage{
summariseData(data, groupBy)
}
\arguments{
\item{groupBy}{is the column user want to summarise by}

\item{df}{is data_total}
}
\value{
A data table that contain the summarise data of cases for the variable that user
choose to summarise.
It will contain 4 columns: One for grouping variable, three for confirmed, estimated
recovered and deceased cases
}
\description{
Summarise data creates a new data frame. It will have one row for each combination
of grouping variables.
}
\examples{
summariseData(data_total, date))
}
