% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariseData_newdaily.R
\name{summariseData_newdaily}
\alias{summariseData_newdaily}
\alias{SummarizeData_newdaily}
\title{#' SummariseData_newdaily daily new cases}
\usage{
summariseData_newdaily(data)
}
\arguments{
\item{data}{is data_total_newdaily
groupBy is the column user want to}
}
\value{
A data table that contain the summarise data of cases for the variable that user
choose to summarise.
}
\description{
#' SummariseData_newdaily daily new cases
}
