% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariseData.R
\name{summariseData}
\alias{summariseData}
\alias{SummarizeData}
\title{Summarise data to display in a data table}
\usage{
summariseData(df, groupBy)
}
\arguments{
\item{df}{is data_total
groupBy is the column user want to summarise by}
}
\value{
A data table that contain the summarise data of cases for the variable that user
choose to summarise.
}
\description{
Summarise data to display in a data table
}
